Modularize reusable pytorch code (from convnet.py to lib/pytorch.py)
